Update CI pipeline to check executable type
The CI script has been updated to check the type of executable file based on the build OS. An additional environment variable `exe_type` is declared to store the expected file type. The script then checks whether the built executable matches this type. Also, added an explicit check on the size of the executable.